Move regulations file
